2179: create time and attendance entity (#887)
* create time and attendance entity

* create time and attendance entity

+++ src/main/java/shamu/company/employee/dto/CompensationDto.java
@@ -1,5 +1,6 @@
 package shamu.company.employee.dto;
 
+import java.math.BigInteger;
 import lombok.Data;
 
 @Data
@@ -7,7 +8,7 @@ public class CompensationDto {
 
   private String id;
 
-  private Double wage;
+  private BigInteger wage;
 
   private SelectFieldInformationDto compensationFrequency;
 }
