+++ src/test/java/shamu/company/job/service/JobUserServiceTests.java
@@ -1,5 +1,6 @@
 package shamu.company.job.service;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedList;
@@ -125,7 +126,7 @@ class JobUserServiceTests {
     void init() {
       jobUpdateDto.setJobId("1");
       jobUpdateDto.setCompensationFrequencyId("1");
-      jobUpdateDto.setCompensationWage(1000.0);
+      jobUpdateDto.setCompensationWage(BigInteger.valueOf(1000));
       jobUpdateDto.setManagerId("2");
       jobUpdateDto.setStartDate(DateUtil.getToday());
